+++ src/__tests__/snapshots/components/ButtonWithPopover.test.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import renderer from 'react-test-renderer';
-import { ButtonWithPopover } from '../..';
-import { DataInformation } from '../../types/dataInformation';
+import { ButtonWithPopover } from '../../..';
+import { DataInformation } from '../../../types/dataInformation';
 
 it('renders ButtonWithPopover with more info button correctly', () => {
     const fetchDataToShow = jest.fn();
